requirements.txt
tensorflow==2.17.0
streamlit==1.32.0
numpy>=1.26.4
pillow==10.3.0